feat: add EXPLORE_AFTER_DEATH build target
This commit grants players the choice to enter explore mode when their
character is about to die. To enable this choice, compile with
EXPLORE_AFTER_DEATH=y

This is not particularly well-tested. In particular, HUPs during the prompt
*should* proceed with silently killing the character and generating a
morgue as normal, but I'm finding it hard to convince myself I haven't
missed any hidden dragons. Weird bugs may occur; please don't bother the
devteam with any bugs you may find in this build.

For the record, I'm not in favor of merging this feature into vanilla DCSS,
but here it is for people to play with, if they want.

(cherry picked from commit d058d7d33c4d6a87895776ab40370a315d2c69de)